// src/features/property/admin/constants/enums.ts
export enum PROPERTY_TYPE {
    VENTA = 'SALE',
    ALQUITLER = 'RENT'
}

export enum CURRENCY {
    PEN = 'PEN',
    USD = 'USD'
}

export enum PROPERTY_CATEGORY {
    CASA = 'HOUSE',
    APARTAMENTO = 'APARTMENT',
    TERRENO = 'LAND',
    COMERCIAL = 'COMMERCIAL',
    OFICINA = 'OFFICE',
    ALMACEN = 'WAREHOUSE',
}